refactor: split code & main entry point
* move anime-dl buisness logic to lib/anime-dl
* extract helpers
