+++ lib/utils/http.js
@@ -1,0 +1,17 @@
+const https = require("https");
+
+const httpGet = (url) => {
+  return new Promise((resolve, reject) => {
+    https
+      .get(url, (res) => {
+        let data = "";
+        res.on("data", (chunk) => (data += chunk));
+        res.on("end", () => resolve(data));
+      })
+      .on("error", reject);
+  });
+};
+
+module.exports = {
+  httpGet,
+};
