+++ lib/utils/index.js
@@ -1,0 +1,7 @@
+const general = require("./general");
+const http = require("./http");
+
+module.exports = {
+  general,
+  http,
+};
